Move CollectionLoader from collection.py to loaders/
